backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-auth
pytest
httpx
requests
email-validator
sqlalchemy
alembic
